+++ lib/polist/service/middleware.rb
@@ -12,7 +12,7 @@ class Polist::Service::Middleware
 
   attr_reader :service
 
-  %i[fail! error! success! form form_attributes].each do |service_method|
+  %i[fail! success! form form_attributes].each do |service_method|
     define_method(service_method) do |*args|
       service.send(service_method, *args)
     end